refactor(app): extract useActivityFeedWiring from useAppModel
