+++ src/renderer/shared/hooks/app/useAppModel.ts
@@ -34,6 +34,7 @@ import {
   WATCH_ENGINE_INITIAL_STATE,
   type WatchStallTracker,
 } from "@renderer/shared/hooks/watch";
+import { useActivityFeedWiring } from "./useActivityFeedWiring";
 import { useDebugCpu } from "./useDebugCpu";
 import { useDebugSnapshot } from "./useDebugSnapshot";
 import { isGameActionable, usePriorityOrchestration } from "@renderer/shared/hooks/priority";
@@ -46,8 +47,6 @@ import { canEarnDrop } from "@renderer/shared/domain/inventory";
 import { sameGameName } from "@renderer/shared/domain/gameName";
 import type { FilterKey, View } from "@renderer/shared/types";
 import { isVerboseLoggingEnabled, logDebug, logInfo } from "@renderer/shared/utils/logger";
-import { recordActivity } from "@renderer/shared/utils/activityFeed";
-import type { ActivityEvent } from "@renderer/shared/utils/activityFeed";
 
 const STALL_NO_PROGRESS_WINDOW_MS = 15 * 60_000;
 const STALL_NO_PROGRESS_WINDOW_NEAR_END_MS = 3 * 60_000;
@@ -763,74 +762,13 @@ export function useAppModel() {
     watching,
   });
 
-  // --- Activity feed wiring (Sources 2-5) ---
-
-  // Source 2: Auto-switch — rising-edge detect (null → truthy with new `at`)
-  const prevAutoSwitchRef = useRef<typeof autoSwitchInfo>(null);
-  useEffect(() => {
-    const prev = prevAutoSwitchRef.current;
-    prevAutoSwitchRef.current = autoSwitchInfo;
-    if (!autoSwitchInfo) return;
-    if (prev && prev.at === autoSwitchInfo.at) return;
-    recordActivity({
-      kind: "auto-switch",
-      at: autoSwitchInfo.at ?? Date.now(),
-      fromName: autoSwitchInfo.from?.name ?? "",
-      toName: autoSwitchInfo.to?.name ?? "",
-      reason: autoSwitchInfo.reason ?? "",
-    } as Omit<Extract<ActivityEvent, { kind: "auto-switch" }>, "id">);
-  }, [autoSwitchInfo]);
-
-  // Source 3: New drops added — rising-edge detect on added set size
-  const prevAddedRef = useRef<Set<string> | undefined>(undefined);
-  useEffect(() => {
-    const prevSize = prevAddedRef.current?.size ?? 0;
-    const currSize = inventoryChanges?.added?.size ?? 0;
-    prevAddedRef.current = inventoryChanges?.added;
-    if (currSize > prevSize && currSize > 0) {
-      const firstAddedId = inventoryChanges.added.values().next().value;
-      const sample = firstAddedId ? inventoryItems.find((it) => it.id === firstAddedId) : undefined;
-      recordActivity({
-        kind: "new-drops",
-        at: Date.now(),
-        count: currSize,
-        sampleTitle: sample?.title,
-      } as Omit<Extract<ActivityEvent, { kind: "new-drops" }>, "id">);
-    }
-  }, [inventoryChanges, inventoryItems]);
-
-  // Source 4: Watch error — rising-edge detect (new or changed error)
-  const prevWatchErrorRef = useRef<typeof watchStats.lastError>(null);
-  useEffect(() => {
-    const prev = prevWatchErrorRef.current;
-    const curr = watchStats.lastError;
-    prevWatchErrorRef.current = curr;
-    if (!curr) return;
-    if (prev && prev.code === curr.code && prev.message === curr.message) return;
-    recordActivity({
-      kind: "watch-error",
-      at: Date.now(),
-      message: curr.message,
-      code: curr.code,
-    } as Omit<Extract<ActivityEvent, { kind: "watch-error" }>, "id">);
-  }, [watchStats.lastError]);
-
-  // Source 5: Watch started — null → truthy transition
-  const prevWatchingRef = useRef<typeof watching>(null);
-  useEffect(() => {
-    const prev = prevWatchingRef.current;
-    prevWatchingRef.current = watching;
-    if (!prev && watching) {
-      recordActivity({
-        kind: "watch-started",
-        at: Date.now(),
-        channelName: watching.name ?? watching.login ?? "",
-        game: watching.game,
-      } as Omit<Extract<ActivityEvent, { kind: "watch-started" }>, "id">);
-    }
-  }, [watching]);
-
-  // --- End activity feed wiring ---
+  useActivityFeedWiring({
+    autoSwitchInfo,
+    inventoryChanges,
+    inventoryItems,
+    lastWatchError: watchStats.lastError,
+    watching,
+  });
 
   useClaimProbe({
     watching,
